requirements.txt
langgraph>=1.0.8
langchain>=0.3.0
langchain-core>=1.2.13
langchain-google-genai>=4.2.0
langfuse>=3.14.3
mlflow>=3.9.0
pandas>=2.3.3
pydantic>=2.12.5
python-dotenv>=1.2.1
